Avoid locking files being run
This is particularly an issue on Windows (I couldn't reproduce it
on Ubuntu) and is somewhat annoying. The main issue is that when
combined with the PyScripter we ship the files remain locked even
after the simulator seems to have been closed. This is an artefact
of the way that PyScripter handles its debuggee processes (it keeps
a single persistent debuggee interpreter around until you either
crash it or force a re-init).

By reading the files and closing them upfront we avoid locking the
user's files, even though their code can still end up running.

Change-Id: I508af7a9e63576b0c1e266d715fe4b5208433dad