Also warn PyScripter users about the daemon threads
Most of SR's competitors will be using PyScripter (on Windows) and
are likely to see odd behaviour resulting from it not killing its
daemon threads when the debuggee tries to exit. Emit a warning so
that they have a way to know what they can do about this if it
bothers them.

Change-Id: I9ad2a4b550ccad5b5619fdd89c610ba93e3cb795